tests/requirements.txt
pytest>=7.4.0
requests>=2.31.0
pytest-html>=4.1.0
pytest-xdist>=3.3.0
pytest-timeout>=2.2.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
